requirements.txt
aiogram==3.13.1
python-dotenv==1.0.1
SQLAlchemy==2.0.36
aiosqlite==0.20.0
APScheduler==3.10.4
yookassa==3.3.1
aiohttp==3.9.5
Jinja2==3.1.4